+++ CHANGELOG.md
@@ -18,8 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The CVSS v2 BaseScore calculator calculates the score on the client side now. [#2536](https://github.com/greenbone/gsa/pull/2536)
 
 ### Fixed
+- Added the deprecatedBy field to CPEs [#2751](https://github.com/greenbone/gsa/pull/2751)
 - Fixed the severity for different advisories [#2611](https://github.com/greenbone/gsa/pull/2611)
-- Added the deprecated_by field to CPEs [#2751](https://github.com/greenbone/gsa/pull/2751)
 
 ### Removed
 - Removed Edge <= 18 support [#2691](https://github.com/greenbone/gsa/pull/2691)